fix: keep git status polling lint-clean
